move scroll.ts to common/controls
