+++ ui/common/src/controls.ts
@@ -50,3 +50,16 @@ export const boolPrefXhrToggle = (prefKey: string, val: boolean, effect: () => v
     await xhr.text(`/pref/${prefKey}`, { method: 'post', body: xhr.form({ [prefKey]: v ? '1' : '0' }) });
     effect();
   });
+
+export function stepwiseScroll(inner: (e: WheelEvent, scroll: boolean) => void): (e: WheelEvent) => void {
+  let scrollTotal = 0;
+  return (e: WheelEvent) => {
+    scrollTotal += e.deltaY * (e.deltaMode ? 40 : 1);
+    if (Math.abs(scrollTotal) >= 4) {
+      inner(e, true);
+      scrollTotal = 0;
+    } else {
+      inner(e, false);
+    }
+  };
+}
