--- ui/common/src/scroll.ts
@@ -1,12 +1,0 @@
-export function stepwiseScroll(inner: (e: WheelEvent, scroll: boolean) => void): (e: WheelEvent) => void {
-  let scrollTotal = 0;
-  return (e: WheelEvent) => {
-    scrollTotal += e.deltaY * (e.deltaMode ? 40 : 1);
-    if (Math.abs(scrollTotal) >= 4) {
-      inner(e, true);
-      scrollTotal = 0;
-    } else {
-      inner(e, false);
-    }
-  };
-}
